backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pgvector==0.2.5
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
pydantic-settings==2.2.1
pydantic==2.7.1
python-multipart==0.0.9
unidecode==1.3.8
passlib==1.7.4
bcrypt==4.0.1
alembic==1.13.1
pandas==2.2.2
openpyxl==3.1.2
requests==2.31.0
pytest==8.1.1
pytest-order==1.2.0
Shapely==2.0.4
geopandas==0.14.3
fiona==1.9.6
folium==0.16.0
openai>=1.50.0
pypdf==4.2.0
google-generativeai==0.5.2
jinja2==3.1.3